Create presentation module to separate unit-testable view codes
