+++ app/build.gradle.kts
@@ -50,6 +50,7 @@ addJUnit()
 dependencies {
   implementation(project(":domain"))
   implementation(project(":data"))
+  implementation(project(":presentation"))
   implementation(project(":core"))
   implementation(project(":login"))
   implementation(project(":playlists"))
