+++ domain/src/main/java/com/github/hadilq/youtubeapp/domain/entity/ConnectionResult.kt
@@ -1,0 +1,9 @@
+package com.github.hadilq.youtubeapp.domain.entity
+
+sealed class ConnectionResult {
+
+  abstract val code: Int
+
+  data class Success(override val code: Int) : ConnectionResult()
+  data class Fail(override val code: Int) : ConnectionResult()
+}
