+++ login/src/main/java/com/github/hadilq/youtubeapp/login/di/LoginModule.kt
@@ -18,7 +18,8 @@ package com.github.hadilq.youtubeapp.login.di
 import com.github.hadilq.youtubeapp.core.di.CoreModule
 import com.github.hadilq.youtubeapp.domain.di.DomainModule
 import com.github.hadilq.youtubeapp.domain.di.LoginModuleSyntax
-import com.github.hadilq.youtubeapp.login.LoginViewModelFactory
+import com.github.hadilq.youtubeapp.presentation.login.LoginViewModelFactory
+import com.github.hadilq.youtubeapp.presentation.di.PresentationModule
 
 interface LoginModule : DomainModule, CoreModule {
 
@@ -26,3 +27,5 @@ interface LoginModule : DomainModule, CoreModule {
 }
 
 fun LoginModuleSyntax.fix() = this as LoginModule
+
+fun LoginModule.fixPresentation() = this as PresentationModule
