+++ presentation/src/main/java/com/github/hadilq/youtubeapp/presentation/di/PresentationModule.kt
@@ -1,0 +1,12 @@
+package com.github.hadilq.youtubeapp.presentation.di
+
+import com.github.hadilq.youtubeapp.domain.di.DomainModule
+import kotlinx.coroutines.CoroutineScope
+
+interface PresentationModule {
+
+  val viewModelScope: CoroutineScope
+}
+
+fun PresentationModule.fixDomain() = this as DomainModule
+
